--- dojo-blog/src/components/PostList.vue
@@ -1,26 +1,0 @@
-<template>
-  <div class="post-list">
-    <div v-for="post in posts" :key="post.id">
-      <SinglePost :post="post" />
-    </div>
-  </div>
-</template>
-
-<script>
-import { onMounted, onUnmounted, onUpdated } from 'vue'
-
-// component imports
-import SinglePost from './SinglePost.vue'
-
-export default {
-  props: ['posts'],
-  components: { SinglePost },
-  setup() {
-
-  }
-}
-</script>
-
-<style>
-
-</style>
