--- dojo-blog/src/composables/getPosts.js
@@ -1,28 +1,0 @@
-import { ref } from 'vue'
-import { projectFirestore } from '../firebase/config'
-
-const getPosts = () => {
-
-  const posts = ref([])
-  const error = ref(null)
-
-  const load = async () => {
-    try {
-      const res = await projectFirestore.collection('posts')
-        //.orderBy('title', 'desc')
-        .orderBy('createdAt', 'desc')  
-        .get()
-
-      posts.value = res.docs.map(doc => {
-        return { ...doc.data(), id: doc.id }
-      })
-    }
-    catch(err) {
-      error.value = err.message
-    }
-  }
-
-  return { posts, error, load }
-}
-
-export default getPosts
