+++ live-chat/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "dojo-blog",
+  "name": "live-chat",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,7 +8,6 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "firebase": "^7.21.1",
     "vue": "^3.0.0-0",
     "vue-router": "^4.0.0-0"
   },
