+++ live-chat/src/main.js
@@ -2,4 +2,7 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 
+// import global css
+import './assets/main.css'
+
 createApp(App).use(router).mount('#app')
