+++ live-chat/src/router/index.js
@@ -1,0 +1,17 @@
+import { createRouter, createWebHistory } from 'vue-router'
+import Welcome from '../views/Welcome.vue'
+
+const routes = [
+  {
+    path: '/',
+    name: 'Welcome',
+    component: Welcome
+  }
+]
+
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
+  routes
+})
+
+export default router
