+++ live-chat/src/views/Welcome.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="welcome container">
+    <p>welcome</p>
+  </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+  .welcome {
+    text-align: center;
+    padding: 20px 0;
+  }
+</style>
